Permission checks added on iterative child deletion and expungment. For deletes, it now halts when reaching a child already having ancestor deleted flag set. Fixed status code to 400 instead of 500 for malformed request. Adding logging statements in moveObject for testing to isolate getObject issue with changes

+++ cmd/metadataconnector/libs/dao/dao.DeleteObject.go
@@ -69,9 +69,22 @@ func (dao *DataAccessLayer) DeleteObject(object *models.ODObject, explicit bool)
 	// Process children
 	resultset, err := dao.GetChildObjects("", 1, 10000, dbObject)
 	for i := 0; i < len(resultset.Objects); i++ {
-		err = dao.DeleteObject(&resultset.Objects[i], false)
-		if err != nil {
-			return err
+		if !resultset.Objects[i].IsAncestorDeleted {
+			authorizedToDelete := false
+			for _, permission := range resultset.Objects[i].Permissions {
+				if permission.Grantee == object.ModifiedBy &&
+					permission.AllowDelete {
+					authorizedToDelete = true
+					break
+				}
+			}
+			if authorizedToDelete {
+				resultset.Objects[i].ModifiedBy = object.ModifiedBy
+				err = dao.DeleteObject(&resultset.Objects[i], false)
+				if err != nil {
+					return err
+				}
+			}
 		}
 	}
